Настроена панель администратора и добавлены функции редактирования и удаления текстов.

+++ SpeedTyping_Angular-ASP.NET/Domain/Repositories/EF/EFTextRepository.cs
@@ -32,6 +32,8 @@ namespace SpeedTyping.Domain.Repositories.EF
         public void DeleteTextByName(string name)
         {
             var textToDelete = context.Texts.FirstOrDefault(x => x.Title == name);
+            if (textToDelete == null)
+                return;
             context.Texts.Remove(textToDelete);
             context.SaveChanges();
         }
